+++ lib/JFFNetwork/DownloadManager/Detail/NSMutableDictionary+DownloadingFileInfo.m
@@ -29,7 +29,8 @@
     {
         return [ fileLength_ unsignedLongLongValue ];
     }
-    return NSURLResponseUnknownLength;
+    
+    return (unsigned long long)NSURLResponseUnknownLength;
 }
 
 +(void)setFileLength:( unsigned long long )fileLength_
